requirements.txt
Django==4.2.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
psycopg2-binary==2.9.6  
django-redis==5.3.0  
djangorestframework-jwt==1.11.0  
